requirements.txt
django
channels
daphne
paho-mqtt
whitenoise
psycopg2-binary
